-lightAtPoint:getter: has much looser restrictions on buffer_element_t.

+++ GutsyStorm/GSNeighborhood.h
@@ -55,7 +55,7 @@ typedef enum
 /* Returns the lighting value at the specified block position for the specified lighting buffer.
  * Assumes the caller is already holding the lock on this buffer on all neighbors.
  */
-- (uint8_t)lightAtPoint:(GSIntegerVector3)p getter:(GSByteBuffer* (^)(GSChunkVoxelData *c))getter;
+- (unsigned)lightAtPoint:(GSIntegerVector3)p getter:(GSByteBuffer* (^)(GSChunkVoxelData *c))getter;
 
 - (BOOL)tryReaderAccessToVoxelDataUsingBlock:(void (^)(void))block;
 - (void)readerAccessToVoxelDataUsingBlock:(void (^)(void))block;
